move effect requirements to separate file
